views_util: Cut import of narrow.
